refactor: adapt files to new next hierarchy
