+++ app.territoiresentransitions.react/app/collectivite/[collectiviteId]/referentiel/[referentielId]/(overview)/@tabs/evolutions/page.tsx
@@ -1,0 +1,5 @@
+import ScoreEvolutions from '@/app/referentiels/evolutions';
+
+export default function Page() {
+  return <ScoreEvolutions />;
+}
